feat(web): update appellant case mapper to handle S20 appeal type
